de-dupe s3 listing logic
