--- empirical-data/common/src/main/scala/com/azavea/ingest/common/HydrateRDDUtils.scala
@@ -1,36 +1,0 @@
-package com.azavea.ingest.common
-
-import org.apache.spark._
-import org.apache.spark.rdd.RDD
-import org.geotools.data.DataStoreFinder
-import org.geotools.data.simple.SimpleFeatureStore
-import org.opengis.feature.simple._
-import org.geotools.feature.simple._
-import org.geotools.feature._
-import com.amazonaws.auth._
-import com.amazonaws.services.s3.AmazonS3Client
-import com.amazonaws.services.s3.model.{ListObjectsRequest, ObjectListing}
-
-import java.util.HashMap
-import scala.collection.JavaConverters._
-import scala.collection.mutable
-
-trait HydrateRDDUtils {
-  val cred = new DefaultAWSCredentialsProviderChain()
-  val client = new AmazonS3Client(cred)
-
-  // Copied from GeoTrellis codebase
-  def listKeys(listObjectsRequest: ListObjectsRequest): Seq[String] = {
-    var listing: ObjectListing = null
-    val result = mutable.ListBuffer[String]()
-    do {
-      listing = client.listObjects(listObjectsRequest)
-      // avoid including "directories" in the input split, can cause 403 errors on GET
-      result ++= listing.getObjectSummaries.asScala.map(_.getKey).filterNot(_ endsWith "/")
-      listObjectsRequest.setMarker(listing.getNextMarker)
-    } while (listing.isTruncated)
-
-    result.toSeq
-  }
-
-}
